clean up and name change

+++ src/socket_handlers/open_orders_socket_handler.py
@@ -13,6 +13,7 @@ class OpenOrdersSocketHandler(SocketHandlerBase):
         self.api_token:   str = api_token
         self.open_orders: dict = { }
         self.open_symbol_pairs = set()
+        self.count: int = 0
 
         self.db = pymongo.MongoClient()[DB.DATABASE_NAME]
         self.collection = self.db[DB.COLLECTION_OO]
@@ -20,35 +21,40 @@ class OpenOrdersSocketHandler(SocketHandlerBase):
 
     def ws_message(self, ws: WebSocketApp, message: str) -> None:
         message = json.loads(message)
-        
-        if isinstance(message, list):
-            for open_orders in message[0]:
-                for txid, order_info in open_orders.items():
-                    if order_info[Status.STATUS] == Status.PENDING or order_info[Status.STATUS] == Status.OPEN:
-                        self.open_orders[txid] = order_info
 
-                        # add symbol_pair to open_symbol_pairs
-                        pair = order_info["descr"]["pair"].split("/")
-                        symbol_pair = pair[0] + pair[1]
-                        self.open_symbol_pairs.add(symbol_pair)
+        if self.count > 3: # at 3, we will only check the new orders, not the current open orders
+            if isinstance(message, list):
+                for open_orders in message[0]:
+                    for txid, order_info in open_orders.items():
+                        if order_info[Status.STATUS] == Status.PENDING or order_info[Status.STATUS] == Status.OPEN:
+                            self.open_orders[txid] = order_info
 
-                        # if its not in the database, add it
-                        if self.collection.count_documents({txid: order_info}) == 0:
-                            self.collection.insert_one({txid: order_info})
+                            # add symbol_pair to open_symbol_pairs
+                            # pair = order_info["descr"]["pair"].split("/")
+                            # symbol_pair = pair[0] + pair[1]
+                            # self.open_symbol_pairs.add(symbol_pair)
 
-                        G.log.pprint_and_log(f"openOrders: open order", order_info, G.lock)
-                    if order_info[Status.STATUS] == Status.CANCELED:
-                        self.open_orders.pop(txid)
-                        self.collection.delete_one({txid: order_info})
+                            # grab the open orders from the data base and store them in self.open_symbol_pairs
 
-                        pair = order_info["descr"]["pair"].split("/")
-                        symbol_pair = pair[0] + pair[1]
-                        if symbol_pair in self.open_symbol_pairs:
-                            self.open_symbol_pairs.remove(symbol_pair)
-                        G.log.pprint_and_log(f"openOrders: canceled order", message, G.lock)
-        else:
-            if "heartbeat" not in message.values():
-                G.log.pprint_and_log("openOrders:", message, G.lock)
+
+                            # if its not in the database, add it
+                            if self.collection.count_documents({txid: order_info}) == 0:
+                                self.collection.insert_one({txid: order_info})
+
+                            G.log.pprint_and_log(f"openOrders: open order", order_info, G.lock)
+                        if order_info[Status.STATUS] == Status.CANCELED:
+                            self.open_orders.pop(txid)
+                            self.collection.delete_one({txid: order_info})
+
+                            pair = order_info["descr"]["pair"].split("/")
+                            symbol_pair = pair[0] + pair[1]
+                            if symbol_pair in self.open_symbol_pairs:
+                                self.open_symbol_pairs.remove(symbol_pair)
+                            G.log.pprint_and_log(f"openOrders: canceled order", message, G.lock)
+            else:
+                if "heartbeat" not in message.values():
+                    G.log.pprint_and_log("openOrders:", message, G.lock)
+        self.count += 1
         return
 
     def ws_open(self, ws: WebSocketApp) -> None:
